Move code from Program Renderer View to ViewModel
